% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_app_prenomUI.R
\name{app_prenomUI}
\alias{app_prenomUI}
\alias{app_prenom}
\title{app_prenomUI}
\usage{
app_prenomUI(id)

app_prenom(input, output, session, all_names)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{all_names}{vector of name to display}
}
\description{
app_prenomUI

app_prenom
}
