% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{draw_a_name}
\alias{draw_a_name}
\title{draw popularity}
\usage{
draw_a_name(data, nom, color)
}
\arguments{
\item{data}{dataset}

\item{nom}{first name to plot}

\item{color}{color to use}
}
\description{
draw popularity
}
